[CIR][ObjC] Add type conversion for Objective-C types
This commit adds support for converting Clang AST Objective-C types to
ClangIR types in the code generator.

Type mappings:
- ObjCObjectPointerType (id) → !cir.objc.id
- ObjCObjectPointerType (Class) → !cir.objc.class
- ObjCObjectPointerType (Foo*) → !cir.objc.interface<"Foo">
- Qualified id (id<Protocol>) → !cir.objc.id (protocol info lost for MVP)

The conversion is added to CIRGenTypes::convertType() in the
Type::ObjCObjectPointer case.

This enables the code generator to handle Objective-C pointer types
in function signatures, local variables, and expressions.

No standalone test for this commit - it will be tested by the
end-to-end codegen tests in the next commit.